// react/typings/ProductType.d.ts
export interface ProductContextState {
    loadingItem: boolean
    selectedItem?: Item | null
    product: MaybeProduct
    selectedQuantity: number
    skuSelector: Partial<SkuSelectorContextState>
    buyButton: BuyButtonContextState
    assemblyOptions: {
      items: Record<GroupId, AssemblyOptionItem[]>
      inputValues: Record<GroupId, InputValues>
      areGroupsValid: Record<GroupId, boolean>
    }
  }
  export interface SkuSelectorContextState {
    selectedImageVariationSKU: string | null
    isVisible: boolean
    areAllVariationsSelected: boolean
  }
  export interface BuyButtonContextState {
    clicked: boolean
  }
type GroupId = string
export interface AssemblyOptionItem {
    id: string
    quantity: number
    seller: string
    initialQuantity: number
    choiceType: string
    name: string
    price: number
    children: Record<string, AssemblyOptionItem[]> | null
  }
  type InputValues = Record<string, string>
type Maybe<T> = T | null | undefined

export type MaybeProduct = Maybe<Product>


export type Product = {
    brand: string
    brandId: string
    cacheId: string
    categories: string[]
    categoriesIds: string[]
    categoryId: string
    categoryTree: Array<{ id: string; name: string; href: string }>
    clusterHighlights: Array<{ id: string; name: string }>
    description: string
    itemMetadata: ItemMetadata
    items: Item[]
    link: string
    linkText: string
    metaTagDescription: string
    priceRange: {
      sellingPrice: { highPrice: number; lowPrice: number }
      listPrice: { highPrice: number; lowPrice: number }
    }
    productClusters: Array<{ id: string; name: string }>
    productId: string
    productName: string
    productReference: string
    properties: ProductSpecification[]
    skuSpecifications: SkuSpecification[]
    specificationGroups: SpecificationGroup[]
    titleTag: string
  }
// ...
export type Item = {
    complementName: string
    ean: string
    images: Array<{
      imageId: string
      imageLabel: string
      imageTag: string
      imageUrl: string
      imageText: string
    }>
    itemId: string
    measurementUnit: string
    name: string
    nameComplete: string
    referenceId: Array<{ Key: string; Value: string }>
    sellers: Seller[]
    unitMultiplier: number
    variations: Array<{ name: string; values: string[] }>
    videos: Array<{ videoUrl: string }>
  }
  export type Seller = {
    sellerId: string
    sellerName: string
    addToCartLink: string
    sellerDefault: boolean
    commertialOffer: CommercialOffer
  }